backend/requirements.txt
Django==3.2
djangorestframework==3.14.0
celery==5.2.7
redis==4.4.0
djangorestframework-simplejwt==5.3.1
requests==2.28.1